requirements.txt
PyYaml
boto3
xmltodict
Jinja2
